+++ lib/protocol/multipart.rb
@@ -4,6 +4,7 @@
 # Copyright, 2025, by Samuel Williams.
 
 require_relative "multipart/version"
+require_relative "multipart/header"
 require_relative "multipart/parser"
 require_relative "multipart/mixed"
 require_relative "multipart/part"