+++ releases.md
@@ -1,3 +1,7 @@
 # Releases
 
+## Unreleased
+
+  - Add `Protocol::Multipart::Header` for parsing parameterized header values.
+
 ## v0.1.0